code reorganization: extracted GeneratedModel into separate files
