+++ src/client/CMakeLists.txt
@@ -27,7 +27,8 @@ pybind11_add_module(${CLIENT_TARGET} MODULE
 		PyCallbacks.cpp
 		PRTContext.cpp
 		PythonLogHandler.cpp
-		InitialShape.cpp)
+		InitialShape.cpp
+		GeneratedModel.cpp)
 
 set_target_properties(${CLIENT_TARGET}
 		PROPERTIES CXX_STANDARD 17
